+++ src/game/tests.rs
@@ -418,6 +418,7 @@ mod old_school_permanents;
 mod old_school_spells;
 mod old_school_upkeep;
 mod poison;
+mod populate;
 mod power_blocking;
 mod premodern_cards;
 mod premodern_lands;